backend/requirements.txt
# Backend dependencies
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
distro==1.9.0
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
exceptiongroup==1.2.2
geographiclib==2.0
geopy==2.4.1
googlemaps==4.10.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
openai==1.59.5
psycopg==3.2.3
psycopg-binary==3.2.3
pydantic==2.10.4
pydantic_core==2.27.2
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0

# Frontend dependencies
# axios==1.7.9
# react==18.3.1
# react-dom==18.3.1
# react-router-dom==7.1.1
# @eslint/js==9.13.0
# @types/react==18.3.11
# @types/react-dom==18.3.1
# @vitejs/plugin-react==4.3.3
# autoprefixer==10.4.20
# eslint==9.13.0
# eslint-plugin-react==7.37.1
# eslint-plugin-react-hooks==5.0.0
# eslint-plugin-react-refresh==0.4.13
# globals==15.11.0
# postcss==8.4.49
# tailwindcss==3.4.17
# vite==5.4.9
